% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{monte_carlo_integrate_cpp}
\alias{monte_carlo_integrate_cpp}
\title{Monte Carlo Integration in C++}
\usage{
monte_carlo_integrate_cpp(f, a, b, n)
}
\arguments{
\item{f}{A function to integrate.}

\item{a}{Lower limit of integration.}

\item{b}{Upper limit of integration.}

\item{n}{Number of random samples.}
}
\value{
Estimated value of the integral.
}
\description{
This function estimates the integral of a given function using the Monte Carlo method in C++.
}
