% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{monte_carlo_integrate_r}
\alias{monte_carlo_integrate_r}
\title{Monte Carlo Integration in R}
\usage{
monte_carlo_integrate_r(f, a, b, n)
}
\arguments{
\item{f}{A function to integrate.}

\item{a}{Lower limit of integration.}

\item{b}{Upper limit of integration.}

\item{n}{Number of random samples.}
}
\value{
Estimated value of the integral.
}
\description{
This function estimates the integral of a given function using the Monte Carlo method in R.
}
\examples{
monte_carlo_integrate_r(function(x) x^2, 0, 1, 1000)
}
